clamav-rest/pyproject.toml
[project]
name = "clamav-rest"
version = "0.1.0"
description = ""
authors = [
    {name = "Python Italia",email = "help@pycon.it"},
]
dependencies = [
    "pyclamd>=0.4.0",
    "fastapi>=0.111.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.3.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.4.8",
]
